[refactor] records_stat 패키지 리팩토링 (#87)
