+++ src/main/java/org/recordy/server/bookmark/controller/BookmarkApi.java
@@ -1,0 +1,56 @@
+package org.recordy.server.bookmark.controller;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import org.recordy.server.common.dto.response.CursorBasePaginatedResponse;
+import org.recordy.server.auth.security.resolver.UserId;
+import org.recordy.server.common.message.ErrorMessage;
+import org.recordy.server.record.controller.dto.response.RecordInfoWithBookmark;
+import org.springframework.data.domain.Slice;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestParam;
+
+@Tag(name = "Bookmark Api")
+public interface BookmarkApi {
+
+    @Operation(
+            summary = "레코드 북마크 및 북마크 해제 API",
+            description = "사용자가 특정 레코드를 북마크하거나 북마크 해제합니다.",
+            responses = {
+                    @ApiResponse(
+                            responseCode = "200",
+                            description = "요청이 성공적으로 처리되었습니다.",
+                            content = @Content
+                    ),
+                    @ApiResponse(
+                            responseCode = "404",
+                            description = "Not Found - 존재하지 않는 사용자 또는 레코드입니다.",
+                            content = @Content(
+                                    mediaType = MediaType.APPLICATION_JSON_VALUE,
+                                    schema = @Schema(
+                                            implementation = ErrorMessage.class
+                                    )
+                            )
+                    ),
+                    @ApiResponse(
+                            responseCode = "500",
+                            description = "Internal Server Error - 서버 내부 오류입니다.",
+                            content = @Content(
+                                    mediaType = MediaType.APPLICATION_JSON_VALUE,
+                                    schema = @Schema(
+                                            implementation = ErrorMessage.class
+                                    )
+                            )
+                    )
+            }
+    )
+    public ResponseEntity<Boolean> bookmark(
+            @UserId Long userId,
+            @PathVariable Long recordId
+    );
+}
