+++ src/main/java/org/recordy/server/bookmark/domain/Bookmark.java
@@ -1,16 +1,16 @@
-package org.recordy.server.record_stat.domain;
+package org.recordy.server.bookmark.domain;
 
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import org.recordy.server.user.domain.User;
 import org.recordy.server.record.domain.Record;
+import org.recordy.server.user.domain.User;
 
 @AllArgsConstructor
 @Builder
 @Getter
-public class View {
+public class Bookmark {
 
     private Long id;
     private User user;
