+++ src/main/java/org/recordy/server/bookmark/repository/impl/BookmarkQueryDslRepository.java
@@ -1,11 +1,11 @@
-package org.recordy.server.record_stat.repository.impl;
+package org.recordy.server.bookmark.repository.impl;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.List;
 
 import lombok.RequiredArgsConstructor;
 import org.recordy.server.common.util.QueryDslUtils;
-import org.recordy.server.record_stat.domain.BookmarkEntity;
+import org.recordy.server.bookmark.domain.BookmarkEntity;
 import org.recordy.server.user.domain.QUserEntity;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
@@ -13,7 +13,7 @@ import org.springframework.data.domain.SliceImpl;
 import org.springframework.stereotype.Repository;
 
 import static org.recordy.server.record.domain.QRecordEntity.recordEntity;
-import static org.recordy.server.record_stat.domain.QBookmarkEntity.bookmarkEntity;
+import static org.recordy.server.bookmark.domain.QBookmarkEntity.bookmarkEntity;
 
 @RequiredArgsConstructor
 @Repository
