+++ src/main/java/org/recordy/server/bookmark/service/impl/BookmarkServiceImpl.java
@@ -1,17 +1,15 @@
-package org.recordy.server.record_stat.service.impl;
+package org.recordy.server.bookmark.service.impl;
 
 import java.util.List;
 import java.util.stream.Collectors;
-
 import lombok.RequiredArgsConstructor;
+import org.recordy.server.bookmark.domain.Bookmark;
+import org.recordy.server.bookmark.repository.BookmarkRepository;
+import org.recordy.server.bookmark.service.BookmarkService;
 import org.recordy.server.common.message.ErrorMessage;
 import org.recordy.server.record.domain.Record;
 import org.recordy.server.record.exception.RecordException;
 import org.recordy.server.record.repository.RecordRepository;
-import org.recordy.server.record_stat.domain.Bookmark;
-import org.recordy.server.record_stat.domain.usecase.Preference;
-import org.recordy.server.record_stat.repository.BookmarkRepository;
-import org.recordy.server.record_stat.service.RecordStatService;
 import org.recordy.server.user.domain.User;
 import org.recordy.server.user.exception.UserException;
 import org.recordy.server.user.repository.UserRepository;
@@ -21,7 +19,7 @@ import org.springframework.stereotype.Service;
 
 @RequiredArgsConstructor
 @Service
-public class RecordStatServiceImpl implements RecordStatService {
+public class BookmarkServiceImpl implements BookmarkService {
 
     private final UserRepository userRepository;
     private final RecordRepository recordRepository;
@@ -48,21 +46,10 @@ public class RecordStatServiceImpl implements RecordStatService {
     }
 
     @Override
-    public Preference getPreference(long userId) {
-        return Preference.of(userId, recordRepository.countAllByUserIdGroupByKeyword(userId));
-    }
-
-
-    @Override
-    public Slice<Record> getBookmarkedRecords(long userId, long cursorId, int size) {
-        return bookmarkRepository.findAllByBookmarksOrderByIdDesc(userId, cursorId, PageRequest.ofSize(size))
-                .map(Bookmark::getRecord);
-    }
-
-    @Override
     public List<Boolean> findBookmarks(long userId, List<Record> records) {
         return records.stream()
                 .map(record -> bookmarkRepository.existsByUserIdAndRecordId(userId, record.getId()))
                 .collect(Collectors.toList());
     }
 }
+
