+++ src/main/java/org/recordy/server/record/service/impl/RecordServiceImpl.java
@@ -6,6 +6,8 @@ import java.util.Optional;
 import java.util.Random;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
+import org.recordy.server.bookmark.domain.Bookmark;
+import org.recordy.server.bookmark.repository.BookmarkRepository;
 import org.recordy.server.common.message.ErrorMessage;
 import org.recordy.server.keyword.domain.Keyword;
 import org.recordy.server.record.domain.File;
@@ -15,11 +17,9 @@ import org.recordy.server.record.exception.RecordException;
 import org.recordy.server.record.repository.RecordRepository;
 import org.recordy.server.record.service.FileService;
 import org.recordy.server.record.service.RecordService;
-import org.recordy.server.record.service.S3Service;
 import org.recordy.server.record.service.dto.FileUrl;
-import org.recordy.server.record_stat.domain.View;
-import org.recordy.server.record_stat.repository.ViewRepository;
-import org.recordy.server.record_stat.service.RecordStatService;
+import org.recordy.server.view.domain.View;
+import org.recordy.server.view.repository.ViewRepository;
 import org.recordy.server.user.domain.User;
 import org.recordy.server.user.exception.UserException;
 import org.recordy.server.user.service.UserService;
@@ -36,6 +36,7 @@ public class RecordServiceImpl implements RecordService {
 
     private final RecordRepository recordRepository;
     private final ViewRepository viewRepository;
+    private final BookmarkRepository bookmarkRepository;
     private final FileService fileService;
     private final UserService userService;
 
@@ -114,6 +115,11 @@ public class RecordServiceImpl implements RecordService {
 
     private Slice<Record> getRecentRecordsWithKeywords(List<Keyword> keywords, long cursorId, int size) {
         return recordRepository.findAllByIdAfterAndKeywordsOrderByIdDesc(keywords, cursorId, PageRequest.ofSize(size));
+    }
+    @Override
+    public Slice<Record> getBookmarkedRecords(long userId, long cursorId, int size) {
+        return bookmarkRepository.findAllByBookmarksOrderByIdDesc(userId, cursorId, PageRequest.ofSize(size))
+                .map(Bookmark::getRecord);
     }
 
     @Override
