--- src/main/java/org/recordy/server/record_stat/service/RecordStatService.java
@@ -1,18 +1,0 @@
-package org.recordy.server.record_stat.service;
-
-import java.util.List;
-
-import org.recordy.server.record.domain.Record;
-import org.recordy.server.record_stat.domain.usecase.Preference;
-import org.springframework.data.domain.Slice;
-
-public interface RecordStatService {
-
-    // command
-    boolean bookmark(long userId, long recordId);
-
-    // query
-    Preference getPreference(long userId);
-    Slice<Record> getBookmarkedRecords(long userId, long cursorId, int size);
-    List<Boolean> findBookmarks(long userId, List<Record> records);
-}
