+++ src/main/java/org/recordy/server/view/repository/ViewRepository.java
@@ -1,0 +1,11 @@
+package org.recordy.server.view.repository;
+
+import org.recordy.server.view.domain.View;
+
+public interface ViewRepository {
+
+    // command
+    View save(View view);
+
+    // query
+}
