+++ src/main/java/org/recordy/server/view/repository/impl/ViewJpaRepository.java
@@ -1,6 +1,6 @@
-package org.recordy.server.record_stat.repository.impl;
+package org.recordy.server.view.repository.impl;
 
-import org.recordy.server.record_stat.domain.ViewEntity;
+import org.recordy.server.view.domain.ViewEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ViewJpaRepository extends JpaRepository<ViewEntity, Long> {
