+++ src/main/java/org/recordy/server/view/repository/impl/ViewQueryDslRepository.java
@@ -1,0 +1,12 @@
+package org.recordy.server.view.repository.impl;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+@RequiredArgsConstructor
+@Repository
+public class ViewQueryDslRepository {
+
+    private final JPAQueryFactory jpaQueryFactory;
+}
