+++ src/test/java/org/recordy/server/record/repository/RecordRepositoryIntegrationTest.java
@@ -11,14 +11,14 @@ import org.recordy.server.record.domain.RecordEntity;
 import org.recordy.server.record.domain.UploadEntity;
 import org.recordy.server.record.repository.impl.UploadJpaRepository;
 import org.recordy.server.record.service.dto.FileUrl;
-import org.recordy.server.record_stat.domain.Bookmark;
-import org.recordy.server.record_stat.domain.BookmarkEntity;
-import org.recordy.server.record_stat.domain.View;
-import org.recordy.server.record_stat.domain.ViewEntity;
-import org.recordy.server.record_stat.repository.BookmarkRepository;
-import org.recordy.server.record_stat.repository.ViewRepository;
-import org.recordy.server.record_stat.repository.impl.BookmarkJpaRepository;
-import org.recordy.server.record_stat.repository.impl.ViewJpaRepository;
+import org.recordy.server.bookmark.domain.Bookmark;
+import org.recordy.server.bookmark.domain.BookmarkEntity;
+import org.recordy.server.view.domain.View;
+import org.recordy.server.view.domain.ViewEntity;
+import org.recordy.server.bookmark.repository.BookmarkRepository;
+import org.recordy.server.view.repository.ViewRepository;
+import org.recordy.server.bookmark.repository.impl.BookmarkJpaRepository;
+import org.recordy.server.view.repository.impl.ViewJpaRepository;
 import org.recordy.server.subscribe.domain.Subscribe;
 import org.recordy.server.subscribe.repository.SubscribeRepository;
 import org.recordy.server.user.domain.UserStatus;
