+++ src/test/java/org/recordy/server/record_stat/domain/BookmarkEntityTest.java
@@ -1,6 +1,8 @@
 package org.recordy.server.record_stat.domain;
 
 import org.junit.jupiter.api.Test;
+import org.recordy.server.bookmark.domain.Bookmark;
+import org.recordy.server.bookmark.domain.BookmarkEntity;
 import org.recordy.server.user.domain.UserStatus;
 import org.recordy.server.util.DomainFixture;
