+++ src/test/java/org/recordy/server/record_stat/service/RecordStatServiceTest.java
@@ -6,26 +6,26 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.recordy.server.bookmark.service.BookmarkService;
 import org.recordy.server.mock.FakeContainer;
-import org.recordy.server.record.controller.dto.response.RecordInfoWithBookmark;
 import org.recordy.server.record.domain.Record;
 import org.recordy.server.record.repository.RecordRepository;
-import org.recordy.server.record_stat.domain.Bookmark;
-import org.recordy.server.record_stat.repository.BookmarkRepository;
-import org.recordy.server.user.domain.UserStatus;
+import org.recordy.server.bookmark.repository.BookmarkRepository;
+import org.recordy.server.record.service.RecordService;
 import org.recordy.server.user.repository.UserRepository;
 import org.recordy.server.util.DomainFixture;
 import org.springframework.data.domain.Slice;
 
 public class RecordStatServiceTest {
 
-    private RecordStatService recordStatService;
+    private RecordService recordService;
+    private BookmarkService bookmarkService;
     private BookmarkRepository bookmarkRepository;
 
     @BeforeEach
     void init() {
         FakeContainer fakeContainer = new FakeContainer();
-        recordStatService = fakeContainer.recordStatService;
+        bookmarkService = fakeContainer.bookmarkService;
         bookmarkRepository = fakeContainer.bookmarkRepository;
         UserRepository userRepository = fakeContainer.userRepository;
         RecordRepository recordRepository = fakeContainer.recordRepository;
@@ -40,7 +40,7 @@ public class RecordStatServiceTest {
     @Test
     void bookmark을_통해_북마크를_생성하고_true를_반환받을_수_있다() {
         // given, when
-        boolean result = recordStatService.bookmark(DomainFixture.USER_ID, DomainFixture.RECORD_ID);
+        boolean result = bookmarkService.bookmark(DomainFixture.USER_ID, DomainFixture.RECORD_ID);
 
         // then
         assertAll(
@@ -52,10 +52,10 @@ public class RecordStatServiceTest {
     @Test
     void bookmark을_통해_북마크를_해제하고_false를_반환받을_수_있다() {
         // given
-        recordStatService.bookmark(DomainFixture.USER_ID, DomainFixture.RECORD_ID);
+        bookmarkService.bookmark(DomainFixture.USER_ID, DomainFixture.RECORD_ID);
 
         // when
-        boolean result = recordStatService.bookmark(DomainFixture.USER_ID, DomainFixture.RECORD_ID);
+        boolean result = bookmarkService.bookmark(DomainFixture.USER_ID, DomainFixture.RECORD_ID);
 
         // then
         assertAll(
@@ -67,15 +67,15 @@ public class RecordStatServiceTest {
     @Test
     void getBookmarkedRecords를_통해_커서_이후의_북마크된_레코드를_최신_순서로_읽을_수_있다() {
         // given
-        recordStatService.bookmark(1, 1);
-        recordStatService.bookmark(2, 1);
-        recordStatService.bookmark(1, 2);
-        recordStatService.bookmark(2, 2);
-        recordStatService.bookmark(1, 3);
-        recordStatService.bookmark(2, 3);
+        bookmarkService.bookmark(1, 1);
+        bookmarkService.bookmark(2, 1);
+        bookmarkService.bookmark(1, 2);
+        bookmarkService.bookmark(2, 2);
+        bookmarkService.bookmark(1, 3);
+        bookmarkService.bookmark(2, 3);
 
         // when
-        Slice<Record> result = recordStatService.getBookmarkedRecords(1, 7, 10);
+        Slice<Record> result = recordService.getBookmarkedRecords(1, 7, 10);
 
         // then
         assertAll(
@@ -90,15 +90,15 @@ public class RecordStatServiceTest {
     @Test
     void getBookmarkedRecords를_통해_커서가_제일_오래된_값이라면_아무것도_반환되지_않는다() {
         // given
-        recordStatService.bookmark(1, 1);
-        recordStatService.bookmark(2, 1);
-        recordStatService.bookmark(1, 2);
-        recordStatService.bookmark(2, 2);
-        recordStatService.bookmark(1, 3);
-        recordStatService.bookmark(2, 3);
+        bookmarkService.bookmark(1, 1);
+        bookmarkService.bookmark(2, 1);
+        bookmarkService.bookmark(1, 2);
+        bookmarkService.bookmark(2, 2);
+        bookmarkService.bookmark(1, 3);
+        bookmarkService.bookmark(2, 3);
 
         // when
-        Slice<Record> result = recordStatService.getBookmarkedRecords(1, 1, 10);
+        Slice<Record> result = recordService.getBookmarkedRecords(1, 1, 10);
 
         // then
         assertAll(
