+++ src/test/java/org/recordy/server/util/DomainFixture.java
@@ -10,7 +10,7 @@ import org.recordy.server.record.domain.Record;
 import org.recordy.server.record.domain.RecordEntity;
 import org.recordy.server.record.domain.usecase.RecordCreate;
 import org.recordy.server.record.service.dto.FileUrl;
-import org.recordy.server.record_stat.domain.Bookmark;
+import org.recordy.server.bookmark.domain.Bookmark;
 import org.recordy.server.user.controller.dto.request.TermsAgreement;
 import org.recordy.server.user.domain.usecase.UserSignIn;
 import org.recordy.server.user.domain.User;
